Add function that gives a result based on player and computer choice
Creating a function that check whether choose of
computer and user is rock, paper, scissors, based
on the previous function. the result will give a
message depending on who is winning
based on the previous function.

The function of the computer choice and user choice
are moved into a const so the value will remain
the same in the last function.

Several log to show different results regarding the
choices and results.